backend/flask-service/requirements.txt
# Flask and extensions
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
Flask-WTF==1.1.1
Werkzeug==2.2.3

# Database
psycopg2-binary==2.9.7
SQLAlchemy==1.4.41

# Authentication
PyJWT==2.8.0
bcrypt==4.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.0
pydantic>=2.5.2
marshmallow==3.20.1
APScheduler==3.10.4
psutil==5.9.5
gunicorn==21.2.0
bleach==6.0.0

# Monitoring and logging
sentry-sdk[flask]==1.28.1
structlog==23.1.0
logging-formatter-anticrlf==1.2.1
python-json-logger==2.0.7

# Service discovery and secrets
python-consul==1.1.0
kubernetes==28.1.0
hvac==1.1.1
boto3==1.29.6

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.3.0
responses==0.23.3

# Local shared package - installed via Dockerfile
# -e ../../meeting_shared 